# docs/spec/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "flaps HTTP API",
    "version": "0.1.0",
    "description": "Admin CRUD, SDK key management, SDK identity, OFREP evaluation and ruleset sync surface exposed by flaps-server. Generated by hand from build_router in crates/flaps-server/src/lib.rs (main@b4e15b1); see docs/spec/api-v1.md for the accompanying guide (authentication model, SSE contract, ETag semantics, error format)."
  },
  "components": {
    "securitySchemes": {
      "adminSession": {
        "type": "http",
        "scheme": "bearer",
        "description": "Session bearer token minted by POST /login. Required by every /projects/** admin route (project, environment, flag, segment, flag-env-config and SDK key management)."
      },
      "sdkKey": {
        "type": "http",
        "scheme": "bearer",
        "description": "SDK key bearer token. GET /sdk/whoami and the OFREP evaluation routes accept both server-kind and client-kind keys. The /sync/v1/* routes require a server-kind key; a client-kind key on those routes receives 403 Forbidden (see api-v1.md for the server-vs-client distinction, which this single security scheme does not model)."
      }
    },
...
    "headers": {
...
      "RetryAfterHeader": {
        "description": "Suggested wait time in seconds before retrying, present on 429 responses.",
        "schema": { "type": "integer", "format": "int64" }
      }
    },
    "responses": {
      "Unauthorized": {
        "description": "Missing or invalid authentication credentials.",
        "content": {
          "application/problem+json": { "schema": { "$ref": "#/components/schemas/Problem" } }
        }
      },
      "ForbiddenServerKeyOnly": {
        "description": "The endpoint requires a server-kind SDK key; a client-kind key was supplied.",
        "content": {
          "application/problem+json": { "schema": { "$ref": "#/components/schemas/Problem" } }
        }
      },
...
      "TooManyRequests": {
        "description": "Rate limit exceeded.",
        "headers": {
          "Retry-After": { "$ref": "#/components/headers/RetryAfterHeader" }
        },
        "content": {
          "application/problem+json": { "schema": { "$ref": "#/components/schemas/Problem" } }
        }
      },
      "InternalError": {
        "description": "Unexpected internal error (e.g. a persistence failure).",
        "content": {
          "application/problem+json": { "schema": { "$ref": "#/components/schemas/Problem" } }
        }
      }
    },
    "schemas": {
      "Problem": {
        "type": "object",
        "description": "RFC 9457 problem+json error body.",
        "properties": {
          "type": { "type": "string", "format": "uri", "description": "Stable problem type URI, e.g. https://flaps.dev/problems/not-found." },
          "title": { "type": "string" },
          "status": { "type": "integer" },
          "detail": { "type": "string" }
        },
        "required": ["type", "title", "status", "detail"]
      },
      "ManagedBy": {
        "type": "string",
        "enum": ["local", "federated"],
        "description": "Whether the resource is owned locally or replicated from a federation."
      },
...
      "SdkKeyKind": {
        "type": "string",
        "enum": ["server", "client"],
        "description": "server keys carry full flag data; client keys receive a filtered, client-safe subset."
      },
...
      "Project": {
        "type": "object",
        "properties": {
          "key": { "type": "string" },
          "name": { "type": "string" },
          "description": { "type": ["string", "null"] },
          "external_ref": { "type": ["string", "null"], "description": "Opaque reference to the project in a federated system." },
          "managed_by": { "$ref": "#/components/schemas/ManagedBy" }
        },
        "required": ["key", "name", "description", "external_ref", "managed_by"]
      },
...
      "SdkKeyScope": {
        "type": "object",
        "properties": {
          "project_key": { "type": "string" },
          "environment_key": { "type": "string" }
        },
        "required": ["project_key", "environment_key"]
      },
...
      "WhoamiResponse": {
        "type": "object",
        "properties": {
          "scope": { "$ref": "#/components/schemas/SdkKeyScope" },
          "kind": { "$ref": "#/components/schemas/SdkKeyKind" }
        },
        "required": ["scope", "kind"]
      },
      "LoginRequest": {
        "type": "object",
        "properties": {
          "username": { "type": "string" },
          "password": { "type": "string" }
        },
        "required": ["username", "password"]
      },
      "LoginResponse": {
        "type": "object",
        "properties": {
          "token": { "type": "string", "description": "Opaque bearer token for subsequent Authorization: Bearer headers." },
          "expires_at": { "type": "string", "format": "date-time" }
        },
        "required": ["token", "expires_at"]
      },
...
      "EventPayload": {
        "type": "object",
        "description": "SSE frame payload on GET /sync/v1/events. Intentionally minimal: never carries flag data, only the notification that a new version exists.",
        "properties": {
          "environment": { "type": "string" },
          "version": { "type": "integer", "format": "int64" }
        },
        "required": ["environment", "version"]
      }
    }
  },
  "paths": {
    "/login": {
      "post": {
        "summary": "Verify credentials and mint a session token",
        "operationId": "postLogin",
        "security": [],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/LoginRequest" } } }
        },
        "responses": {
          "200": {
            "description": "Session token minted.",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/LoginResponse" } } }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "429": { "$ref": "#/components/responses/TooManyRequests" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
    "/projects": {
      "get": {
        "summary": "List all projects",
        "operationId": "listProjects",
        "security": [{ "adminSession": [] }],
        "responses": {
          "200": {
            "description": "All projects.",
            "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Project" } } } }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
...
    "/sdk/whoami": {
      "get": {
        "summary": "Identify the calling SDK key and check its rate limit",
        "operationId": "getWhoami",
        "security": [{ "sdkKey": [] }],
        "responses": {
          "200": {
            "description": "Key scope and kind.",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/WhoamiResponse" } } }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "429": { "$ref": "#/components/responses/TooManyRequests" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
...
    "/sync/v1/events": {
      "get": {
        "summary": "Server-sent events stream of ruleset change notifications",
        "description": "Server-key only: a client-kind key receives 403. Bounded by a separate concurrency quota (global and per-key) that rejects excess concurrent subscriptions with 429, distinct from the ordinary request rate limiter. See api-v1.md for the full notify-then-fetch SSE contract and the concurrency quota; text/event-stream is not modeled in detail by OpenAPI.",
        "operationId": "getEvents",
        "security": [{ "sdkKey": [] }],
        "responses": {
          "200": {
            "description": "SSE stream opened. Each frame carries a JSON-encoded EventPayload.",
            "content": { "text/event-stream": { "schema": { "$ref": "#/components/schemas/EventPayload" } } }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/ForbiddenServerKeyOnly" },
          "429": { "$ref": "#/components/responses/TooManyRequests" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    }
  }
}